Creation of the CommonKernelFactory class which help to create few of the most common convolution kernels existing. There is also a demo in the main that create every type of default kernels, apply them and save them to se the results.

+++ KernelProcessing.cpp
@@ -67,7 +67,7 @@ void KernelProcessing::setMaskElement(int row, int column, float value) {
     mask[row * dimension + column] = value;
 }
 
-int KernelProcessing::getMaskElement(int row, int column) const {
+float KernelProcessing::getMaskElement(int row, int column) const {
     if (!itsValidPosition(row, column)) {
         throw std::invalid_argument("Error: given position is not valid.");
     }
